Get doom to start on modern linux via Xephyr
Fixed a bunch of segfaults, typos, and general
errors in the code.

Fun fact, when Doom was created, it was common for displays to
only support psuedocolor with 8 bit color depth. In other words,
X11 could only display 256 colors at a given time.
Displays don't really support pseudocolor anymore(including the monitor
I'm using lol) and use TrueColor/DirectColor when using X11.
This causes an error during startup because linux doom refuses
to startup unless it can use a pseudocolor display.
What we need to do for now is emulate a similar
window as back then. Xephyr is the way to do this.
Xephyr is an implementation of the X11 server that can
emulate a window we need. So for now we emulate and
change slowly.

To run doom, create a linux folder under linuxdoom.
Next run make in the linuxdoom folder.
Next open up a  shell.
Run:
        setup-xephyr.sh &
Then run:
        start.sh
Doom should start up(Although very small, yay
90s resolutions)