Fix scaling of DOOM in Xephyr(kinda)
Was able to fix how small doom was on startup.
The game binary allows optional parameters to
be passed in with the values from -1 to -4.
This controls the multiply variable in i_video.c,
which scales the game in the window.
Set the start script to pass in -3 as a
parameter(since -4 is broken and was never fixed)

Now we can finally play the game somewhat comfortably
(No sound for now)