Get sndserver sound effects to work for linux
The linux port of doom  had a seperate sound server that
handled sounds/music. This is located in the sndserv folder.

All I had to do to fix the sound server to work on modern linux
was include errno.h in linux.c and it built fine.
To have doom use the sndserver, it must be placed in the same
location your wad files are. (Either where the executable is run
or by the DOOMWADDIR environment variable)

When doom was created, linux/unix used /dev/dsp as the main audio device.
This was part of the Open Sound System API. However, due to the project
maintainer of OSS locking features(such as support for new sound
devices) behind a proprietary license, linux moved away from
OSS to ALSA(Advanced Linux Sound Architecture). From there
high level sound servers like PulseAudio were built on top
of ALSA. What this means is that the logic to send
sound to /dev/dsp will no longer work. To get around this until
I rewrite the sound code, I am using padsp(PulseAudio OSS Wrapper)
This will translate any calls to an OSS specific device to
a PulseAudio equivalent. To enable this I must pass the
linuxdoom binary to padsp as an argument. This change was
made in start.sh. There should be equivalent wrappers
for other high level sound APIs(Pipewire, etc) if
others feel the need to run doom using that

Finally, the game was failing to start the sndserver when
it was looking for it in the same directory as
the doom binary. The logic wasn't adding ./ to the sound server
file name so it was failing. Change was made in i_sound.c

THIS ONLY GETS SOUND EFFECTS WORKING. Sadly, the Init_Music
functions in sndserver are empty. Not sure if the linux port
never had music or if it was a copyright issue. Either way,
the music should still exist in the WADs so filling out those
Music functions should restore the music to the game